ENH: Use pyblock to estimate stderr
The function _fp_stderr_catch_errors() previously used the package
flyvbjerg_petersen_std_err, available from
https://github.com/manoharan-lab/flyvbjerg-std-err. The function now defaults
to using the pyblock package, which can be installed with pip.

The pyblock package provides the benefit that, in addition to the stderr, the
error of the stderr will also be estimated. The function now returns the
err_err as the second value of the returned tuple.

In case of exceptions, the function makes two attempts to estimate a stderr,
before it quietly fails by returning np.Inf.