ENH: Calculate std based on stderr of each msd
Each particle has an msd and a stderr calculated using the Flyvbjerg
Petersen Standard Error algorithm. This commit takes that stderr from
each individual msd and calculates a standard deviation of the ensemble
msd using reliability weights that are the inverse square of the
particle msd stderr.

There is one odd thing: the ensemble msd (the emsd) is calculated using
weights based on the frequency, whereas the standard deviation is
calculated using weights based on the stderr.